[mutable] Undo layout id changes because they break some assets. #jira UE-209986 #rnx [FYI] [at]pere.rifa
[CL 33508521 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableTools/Private/MuT/ASTOpImageRasterMesh.h
@@ -24,7 +24,7 @@ namespace mu
 		ASTChild projector;
 
 		/** If layouts are used, this can indicate a single layout block that we want to raster. */
-		uint64 BlockId;
+		int32 BlockId;
 		int8 LayoutIndex;
 
 		/** Size of the image to generate by rasterization of the mesh. */
